-- Login and register was created

+++ src/main/java/com/azericard/insurance/service/CompanyService.java
@@ -2,6 +2,8 @@ package com.azericard.insurance.service;
 
 import com.azericard.insurance.data.CompanyRepository;
 import com.azericard.insurance.entity.Company;
+import com.azericard.insurance.entity.EncodedRole;
+import com.azericard.insurance.exception.AccessNotAllowedException;
 import com.azericard.insurance.exception.CompanyNotFoundException;
 import com.azericard.insurance.exception.GeneralException;
 import org.springframework.stereotype.Service;
@@ -17,25 +19,39 @@ public class CompanyService {
         this.companyRepo = companyRepo;
     }
 
-    public List<Company> getAll() {
-        List<Company> companies = (List<Company>) companyRepo.findAll();
-        if (companies.isEmpty()) {
-            throw new GeneralException("No Data Found");
+    public List<Company> getAll(String role) {
+        if (role.equals(EncodedRole.SUPER_ADMIN)) {
+            List<Company> companies = (List<Company>) companyRepo.findAll();
+            if (companies.isEmpty()) {
+                throw new GeneralException("No Data Found");
+            }
+            return companies;
         }
-        return companies;
+        throw new AccessNotAllowedException();
     }
 
-    public Company getOne(long id) {
-        Optional<Company> optionalCompany = companyRepo.findById(id);
-        return optionalCompany.orElseThrow(CompanyNotFoundException::new);
+    public Company getOne(long id,String role) {
+        if (role.equals(EncodedRole.SUPER_ADMIN)) {
+            Optional<Company> optionalCompany = companyRepo.findById(id);
+            return optionalCompany.orElseThrow(CompanyNotFoundException::new);
+        }
+        throw new AccessNotAllowedException();
+
     }
 
-    public Company save(Company company) {
-        return companyRepo.save(company);
+    public Company save(Company company,String role) {
+        if (role.equals(EncodedRole.SUPER_ADMIN)) {
+            return companyRepo.save(company);
+        }
+        throw new AccessNotAllowedException();
+
     }
 
-    public void delete(Company company) {
-        companyRepo.delete(company);
+    public void delete(Company company,String role) {
+        if (role.equals(EncodedRole.SUPER_ADMIN)) {
+            companyRepo.delete(company);
+        }
+        throw new AccessNotAllowedException();
     }
